refactor(auth): implement acces token on login function

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Classes\ApiResponseClass;
 use App\Http\Controllers\api\AuthController;
 use App\Http\Controllers\api\BeritaController;
 use App\Models\Berita;
@@ -13,7 +14,12 @@ Route::get('/user', function (Request $request) {
 Route::get('/berita', [BeritaController::class, 'getAll']);
 Route::get('/berita/{id}', [BeritaController::class, 'getById']);
 Route::post('/berita', [BeritaController::class, 'store']);
-Route::put('/berita/{id}', [BeritaController::class, 'update']);
-Route::delete('/berita/{id}', [BeritaController::class, 'destroy']);
+Route::put('/berita/{id}', [BeritaController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/berita/{id}', [BeritaController::class, 'destroy'])->middleware('auth:sanctum');
 
-Route::post('/login', [AuthController::class, 'check_login']);
+Route::post('/login', [AuthController::class, 'login']);
+
+//fallback route for unauthorized user
+Route::get('/unauthorize', function () {
+    return ApiResponseClass::sendError('Unauthorized', 403);
+})->name('login');
